resource table rename
