+++ lambdas/lib/dynamo/__mocks__/operations.js
