+++ test/actor/wharfie-deployment.test.js
@@ -132,8 +132,8 @@ describe('deployment IaC', () => {
                 "LOGGING_LEVEL": "info",
                 "MAX_QUERIES_PER_ACTION": "10000",
                 "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                "OPERATIONS_TABLE": "test-deployment-operations",
                 "RESOURCE_QUERY_CONCURRENCY": "10",
-                "RESOURCE_TABLE": "test-deployment-resource",
                 "SEMAPHORE_TABLE": "test-deployment-semaphore",
                 "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                 "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -173,8 +173,8 @@ describe('deployment IaC', () => {
                     "LOGGING_LEVEL": "info",
                     "MAX_QUERIES_PER_ACTION": "10000",
                     "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                    "OPERATIONS_TABLE": "test-deployment-operations",
                     "RESOURCE_QUERY_CONCURRENCY": "10",
-                    "RESOURCE_TABLE": "test-deployment-resource",
                     "SEMAPHORE_TABLE": "test-deployment-semaphore",
                     "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                     "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -308,8 +308,8 @@ describe('deployment IaC', () => {
                           "MAX_QUERIES_PER_ACTION": "10000",
                           "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
                           "NODE_OPTIONS": "--enable-source-maps",
+                          "OPERATIONS_TABLE": "test-deployment-operations",
                           "RESOURCE_QUERY_CONCURRENCY": "10",
-                          "RESOURCE_TABLE": "test-deployment-resource",
                           "SEMAPHORE_TABLE": "test-deployment-semaphore",
                           "STACK_NAME": "test-deployment",
                           "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
@@ -485,8 +485,8 @@ describe('deployment IaC', () => {
                 "LOGGING_LEVEL": "info",
                 "MAX_QUERIES_PER_ACTION": "10000",
                 "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                "OPERATIONS_TABLE": "test-deployment-operations",
                 "RESOURCE_QUERY_CONCURRENCY": "10",
-                "RESOURCE_TABLE": "test-deployment-resource",
                 "SEMAPHORE_TABLE": "test-deployment-semaphore",
                 "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                 "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -526,8 +526,8 @@ describe('deployment IaC', () => {
                     "LOGGING_LEVEL": "info",
                     "MAX_QUERIES_PER_ACTION": "10000",
                     "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                    "OPERATIONS_TABLE": "test-deployment-operations",
                     "RESOURCE_QUERY_CONCURRENCY": "10",
-                    "RESOURCE_TABLE": "test-deployment-resource",
                     "SEMAPHORE_TABLE": "test-deployment-semaphore",
                     "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                     "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -661,8 +661,8 @@ describe('deployment IaC', () => {
                           "MAX_QUERIES_PER_ACTION": "10000",
                           "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
                           "NODE_OPTIONS": "--enable-source-maps",
+                          "OPERATIONS_TABLE": "test-deployment-operations",
                           "RESOURCE_QUERY_CONCURRENCY": "10",
-                          "RESOURCE_TABLE": "test-deployment-resource",
                           "SEMAPHORE_TABLE": "test-deployment-semaphore",
                           "STACK_NAME": "test-deployment",
                           "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
@@ -838,8 +838,8 @@ describe('deployment IaC', () => {
                 "LOGGING_LEVEL": "info",
                 "MAX_QUERIES_PER_ACTION": "10000",
                 "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                "OPERATIONS_TABLE": "test-deployment-operations",
                 "RESOURCE_QUERY_CONCURRENCY": "10",
-                "RESOURCE_TABLE": "test-deployment-resource",
                 "SEMAPHORE_TABLE": "test-deployment-semaphore",
                 "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                 "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -879,8 +879,8 @@ describe('deployment IaC', () => {
                     "LOGGING_LEVEL": "info",
                     "MAX_QUERIES_PER_ACTION": "10000",
                     "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                    "OPERATIONS_TABLE": "test-deployment-operations",
                     "RESOURCE_QUERY_CONCURRENCY": "10",
-                    "RESOURCE_TABLE": "test-deployment-resource",
                     "SEMAPHORE_TABLE": "test-deployment-semaphore",
                     "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                     "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -1014,8 +1014,8 @@ describe('deployment IaC', () => {
                           "MAX_QUERIES_PER_ACTION": "10000",
                           "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
                           "NODE_OPTIONS": "--enable-source-maps",
+                          "OPERATIONS_TABLE": "test-deployment-operations",
                           "RESOURCE_QUERY_CONCURRENCY": "10",
-                          "RESOURCE_TABLE": "test-deployment-resource",
                           "SEMAPHORE_TABLE": "test-deployment-semaphore",
                           "STACK_NAME": "test-deployment",
                           "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
@@ -1191,8 +1191,8 @@ describe('deployment IaC', () => {
                 "LOGGING_LEVEL": "info",
                 "MAX_QUERIES_PER_ACTION": "10000",
                 "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                "OPERATIONS_TABLE": "test-deployment-operations",
                 "RESOURCE_QUERY_CONCURRENCY": "10",
-                "RESOURCE_TABLE": "test-deployment-resource",
                 "SEMAPHORE_TABLE": "test-deployment-semaphore",
                 "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                 "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -1232,8 +1232,8 @@ describe('deployment IaC', () => {
                     "LOGGING_LEVEL": "info",
                     "MAX_QUERIES_PER_ACTION": "10000",
                     "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
+                    "OPERATIONS_TABLE": "test-deployment-operations",
                     "RESOURCE_QUERY_CONCURRENCY": "10",
-                    "RESOURCE_TABLE": "test-deployment-resource",
                     "SEMAPHORE_TABLE": "test-deployment-semaphore",
                     "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
                     "WHARFIE_LOGGING_FIREHOSE": "test-deployment-firehose",
@@ -1367,8 +1367,8 @@ describe('deployment IaC', () => {
                           "MAX_QUERIES_PER_ACTION": "10000",
                           "MONITOR_QUEUE_URL": "test-deployment-monitor-queue",
                           "NODE_OPTIONS": "--enable-source-maps",
+                          "OPERATIONS_TABLE": "test-deployment-operations",
                           "RESOURCE_QUERY_CONCURRENCY": "10",
-                          "RESOURCE_TABLE": "test-deployment-resource",
                           "SEMAPHORE_TABLE": "test-deployment-semaphore",
                           "STACK_NAME": "test-deployment",
                           "TEMPORARY_GLUE_DATABASE": "test-deployment-temporary-database",
@@ -1573,7 +1573,7 @@ describe('deployment IaC', () => {
             "resources": {
               "test-deployment-actor-policy": {
                 "dependsOn": [
-                  "test-deployment-resource",
+                  "test-deployment-operations",
                   "test-deployment-locations",
                   "test-deployment-semaphore",
                   "test-deployment-events",
@@ -1643,7 +1643,7 @@ describe('deployment IaC', () => {
                         ],
                         "Effect": "Allow",
                         "Resource": [
-                          "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-resource",
+                          "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-operations",
                           "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-locations",
                           "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-events",
                           "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-semaphore",
@@ -1785,7 +1785,7 @@ describe('deployment IaC', () => {
                 "dependsOn": [
                   "test-deployment-glue-database",
                   "test-deployment-deployment-resources-logging-resource-role",
-                  "test-deployment-resource",
+                  "test-deployment-operations",
                   "test-deployment-locations",
                   "test-deployment-dependencies",
                   "test-deployment-bucket",
@@ -2812,11 +2812,11 @@ describe('deployment IaC', () => {
                 "resourceType": "Table",
                 "status": "STABLE",
               },
-              "test-deployment-resource": {
+              "test-deployment-operations": {
                 "dependsOn": [],
-                "name": "test-deployment-resource",
+                "name": "test-deployment-operations",
                 "properties": {
-                  "arn": "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-resource",
+                  "arn": "arn:aws:dynamodb:us-east-1:123456789012:table/test-deployment-operations",
                   "attributeDefinitions": [
                     {
                       "AttributeName": "resource_id",
