+++ test/operations/migrate.test.js
@@ -22,7 +22,7 @@ const {
 const { version } = require('../../package.json');
 const daemon_lambda = require('../../lambdas/daemon');
 
-jest.mock('../../lambdas/lib/dynamo/resource');
+jest.mock('../../lambdas/lib/dynamo/operations');
 jest.mock('../../lambdas/lib/dynamo/event');
 jest.mock('../../lambdas/lib/dynamo/location');
 jest.mock('../../lambdas/lib/dynamo/semaphore');
@@ -37,10 +37,10 @@ const { SQS } = require('@aws-sdk/client-sqs');
 const { S3 } = require('@aws-sdk/client-s3');
 const { CloudFormation } = require('@aws-sdk/client-cloudformation');
 
-const resource = require('../../lambdas/lib/dynamo/resource');
+const resource = require('../../lambdas/lib/dynamo/operations');
 const logging = require('../../lambdas/lib/logging');
 
-const dynamo_resource = require('../../lambdas/lib/dynamo/resource');
+const dynamo_resource = require('../../lambdas/lib/dynamo/operations');
 const semaphore = require('../../lambdas/lib/dynamo/semaphore');
 
 const glue = new Glue();
